.........T [ZBXCTR-20] drop params api_server_scheme and api_server_port
We can extract these from api_url, like we do with api_hostname by
adding to the regex. This should also make it less error-prone incase
of missing or inconsistent values in the macros.

We're also shortening the names, to make it feel more in line with
controller, kubelet and scheduler.

Also reusing the code in kube.node.lld.