.........T [ZBXCTR-20] Default KUBE.API.URL to the internal service
localhost:6443 is probably valid in a all-in-one development scenario
with zabbix-server and kubernetes-api all on the same machine. But
the more likely scenario (both in development and production), you're
running these queries via a zabbix-proxy inside your cluster, in which
you'll have to override this.

This url should work out-of-the-box from a zabbix-proxy within the
cluster. Making it one less thing to override in many environments.